+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.1.2] - 2026-01-28
-- fix negative values getWatt, getVWave, getIWave.
-
+- fix negative values getWatt(), getVWave(), getIWave().
+- add **float getMaxCurrent()** **float getMaxCurrentRMS()**
+- add **float getMaxVoltage()** **float getMaxVoltageRMS()**